requirements.txt
pydantic>=2.6.0
feedparser>=6.0.10
web3>=6.0.0
duckdb>=0.10.0
fastapi>=0.110.0
uvicorn>=0.27.0
chromadb>=0.5.0
sentence-transformers>=2.2.2
pysqlite3-binary>=0.5.2
